Create README.md
We will be able to know the interest of customers, Gender wise, age wise and we can know about the annual incomes on the basis of Spending scores, On the basis of these details we can provide more options to the customers. So that they can efficiently take the benefits and buy more and more products.